Added a few extra test chars, and I honestly don't know what I did but now I can build to my phone

+++ Dead Sea iOS/Dead Sea iOS/SpriteNodes/DSCharacterSpriteNode.h
@@ -23,7 +23,7 @@
 /*!
  * @brief Number of shots fired per second
  */
-@property (nonatomic, readwrite) int fireRate;
+@property (nonatomic, readwrite) double fireRate;
 /*!
  * @brief Number of shots that make up one burst
  */
